python/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.0.0
beautifulsoup4>=4.0.0